refactor(test): move fixture to conftest.py for reuse
